fix: revert gfql subdirectory refactoring that caused import conflicts
- Move call_executor.py and call_safelist.py back to compute directory
- Update all imports to use original paths
- Remove empty gfql subdirectory that conflicted with gfql.py
- Fix ASTLet alias and type annotations
- Fix lint issues (newlines, comparison operators)
