+++ graphistry/tests/compute/test_call_operations_gpu.py
@@ -8,7 +8,7 @@ from graphistry.tests.test_compute import CGFull
 from graphistry.Engine import Engine
 from graphistry.compute.ast import ASTCall, ASTQueryDAG, n
 from graphistry.compute.chain_dag import chain_dag_impl
-from graphistry.compute.gfql.call_executor import execute_call
+from graphistry.compute.call_executor import execute_call
 from graphistry.compute.validate_schema import validate_chain_schema
 from graphistry.compute.exceptions import ErrorCode, GFQLTypeError
